@@ -1,6 +1,6 @@
 [metadata]
 name = czitools
-version = 0.10.3
+version = 0.11.0
 author = Sebastian Rhode
 author_email = sebrhode@gmail.com
 url = https://github.com/sebi06/czitools
@@ -16,7 +16,6 @@ classifiers =
     Topic :: Scientific/Engineering
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3.12
@@ -34,7 +33,7 @@ project_urls =
 package_dir =
     = src
 packages = find:
-python_requires = >=3.9, <3.14
+python_requires = >=3.10, <3.14
 
 # add your package requirements here
 install_requires =
