@@ -1,3 +1,3 @@
 # __init__.py
 # version of the czitools package
-__version__ = "0.10.3"
+__version__ = "0.11.0"
